making resolve channel and resolve user with role util function and using in remove leader intent handlers
